Update display-config.php
First time doing this, so my apologies if I'm doing GitHub wrong.  Adding this viewport statement (line 19) will make the page responsive to the screen size, so that it is easily viewable on a smaller screen like a smartphone.  I will propose this change in all of the files I edited on my own Allstar node.